Casey's comments
