--- tests/std/tests/LWG3515_stacktrace_ostream_operator_should_be_less_templatized/env.lst
@@ -1,4 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-RUNALL_INCLUDE ..\usual_latest_matrix.lst
